point 3 import data from csv files

+++ app/Converter/IntegerToRupiah.scala
@@ -5,7 +5,7 @@ import java.util.{Currency, Locale}
 object IntegerToRupiah {
   val rupiahFormatter = java.text.NumberFormat.getCurrencyInstance
   rupiahFormatter.setCurrency(Currency.getInstance(new Locale("id", "ID")))
-  def convert(num:Int):String = {
+  def convert(num:Long):String = {
     val result = rupiahFormatter.format(num)
 
     result.replaceAll("(IDR)","Rp")
